+++ R/etwfe_class.R
@@ -23,6 +23,7 @@ print.etwfe <- function(
 	max_cohorts = getOption("etwfe.max_cohorts", 10),
 	order_by = c("cohort", "estimate", "abs_estimate", "pvalue", "none"),
 	show_internal = FALSE,
+	max_event_times = getOption("etwfe.max_event_times", 10),
 	...
 ) {
 	.print_estimator_output(
@@ -36,6 +37,7 @@ print.etwfe <- function(
 		max_cohorts = max_cohorts,
 		order_by = match.arg(order_by),
 		show_internal = show_internal,
+		max_event_times = max_event_times,
 		...
 	)
 }